+++ injectool/resolvers.py
@@ -1,6 +1,7 @@
 """Dependency resolvers used by container"""
 
 from contextvars import ContextVar, Token
+import threading
 from typing import Any, Callable, Dict, List, Optional, Type
 
 from injectool.core import get_container
@@ -30,7 +31,7 @@ class DependencyScope:
         self._reset_token = _CURRENT_SCOPE.set(self)
         return self
 
-    def __exit__(self, exc_type, exc_val, exc_tb):
+    def __exit__(self, *_):
         """deletes scope as current"""
         if self._resent_token is not None:
             _CURRENT_SCOPE.reset(self._resent_token)
@@ -49,7 +50,7 @@ _CURRENT_SCOPE.set(DependencyScope())
 
 class ScopeResolver:
     """Instance resolver for scope"""
-    def __init__(self, type_, dispose: Optional[Callable[[Any], None]]):
+    def __init__(self, type_: Type, dispose: Optional[Callable[[Any], None]]):
         self._type: Type = type_
         self._dispose: Optional[Callable[[Any], None]] = dispose
         self._instances: Dict[DependencyScope, Any] = {}
@@ -76,3 +77,28 @@ class ScopeResolver:
 def add_scoped(dependency: Any, type_: Type, dispose: Optional[Callable[[Any], None]] = None):
     """Adds type instance per scope to current container"""
     get_container().set(dependency, ScopeResolver(type_, dispose).resolve)
+
+
+class ThreadResolver:
+    """Instance resolver for thread"""
+    def __init__(self, type_: Type):
+        self._type: Type = type_
+        self._instances: Dict[int, Any] = {}
+
+    def resolve(self) -> Any:
+        """returns type instance for current thread"""
+        thread = threading.current_thread()
+        thread_id = threading.get_ident()
+        print(f'id {threading.get_ident()} hash {hash(thread)} native {threading.get_native_id()}')
+
+        if thread_id not in self._instances:
+            instance = self._type()
+            self._instances[thread_id] = instance
+        else:
+            instance = self._instances[thread_id]
+
+        return instance
+
+
+def add_per_thread(dependency: Any, type_: Type):
+    get_container().set(dependency, ThreadResolver(type_).resolve)