+++ injectool/tests/resolvers_tests.py
@@ -1,9 +1,14 @@
+from concurrent.futures.thread import ThreadPoolExecutor
 from unittest.mock import Mock, call
 
 from pytest import mark, fixture
 
 from injectool.core import get_dependency_key, Container, resolve, use_container
-from injectool.resolvers import DependencyScope, add_scoped, add_singleton, add_type
+from injectool.resolvers import DependencyScope, add_per_thread, add_scoped, add_singleton, add_type
+
+
+class SomeClass:
+    pass
 
 
 @fixture
@@ -16,9 +21,10 @@ def container_fixture(request):
 
 
 @mark.parametrize('dependency, value', [
-    ('key', lambda _, __=None: 'value'),
+    ('key', lambda _: 'value'),
     (get_dependency_key, 1),
-    (Container, Mock())
+    (Container, Mock()),
+    (SomeClass, SomeClass())
 ])
 def test_add_singleton(dependency, value):
     """add_singleton() should add SingletonResolver to current container"""
@@ -56,7 +62,7 @@ class ScopesTests:
         (Container, Container)
     ])
     def test_add_scoped(self, dependency, type_):
-        """should uses single instance for container without scope"""
+        """should use single instance for container without scope"""
         add_scoped(dependency, type_)
 
         with DependencyScope():
@@ -70,10 +76,11 @@ class ScopesTests:
 
     @mark.parametrize('dependency, type_', [
         (Mock, Mock),
-        (Container, Container)
+        (Container, Container),
+        (SomeClass, SomeClass)
     ])
     def test_add_scoped_default(self, dependency, type_):
-        """should uses single instance for container without scope"""
+        """should use single instance for container without scope"""
         add_scoped(dependency, type_)
 
         actual = resolve(dependency)
@@ -83,7 +90,8 @@ class ScopesTests:
 
     @mark.parametrize('dependency, type_', [
         (Mock, Mock),
-        (Container, Container)
+        (Container, Container),
+        (SomeClass, SomeClass)
     ])
     def test_add_scoped_dispose(self, dependency, type_):
         """should call dispose when scope is closed"""
@@ -97,4 +105,43 @@ class ScopesTests:
 
         assert dispose.call_args_list[0] == call(actual)
         assert dispose.call_args_list[1] == call(outer)
+
+
+@mark.usefixtures(container_fixture.__name__)
+class ThreadTests:
+    """Thread resolver tests"""
+
+    container: Container
+
+    @mark.parametrize('dependency, type_', [
+        (Mock, Mock),
+        (Container, Container),
+        (SomeClass, SomeClass)
+    ])
+    def test_add_per_thread_instance_per_thread(self, dependency, type_):
+        """should return instance of passed type per thread"""
+        add_per_thread(dependency, type_)
+
+        one = resolve(dependency)
+        two = resolve(dependency)
+
+        assert isinstance(one, type_)
+        assert one is two
+
+
+    @mark.parametrize('dependency, type_', [
+        (Mock, Mock),
+        (Container, Container),
+        (SomeClass, SomeClass)
+    ])
+    def test_add_per_thread_instances(self, dependency, type_):
+        """should return instance per thread"""
+        add_per_thread(dependency, type_)
+        one = resolve(dependency)
+
+        with ThreadPoolExecutor(max_workers=1) as executor:
+            future = executor.submit(self.container.resolve, dependency)
+            two = future.result()
+
+        assert one != two