+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Dependency is used as a key instead of its name
 - Added default injection value
+- Resolvers are removed
 
 ## 2.0.4