+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-# Dev
+# 3.0.0
 
 - Dependency is used as a key instead of its name
 - Added default injection value